Rendering finally clears the transparent layer

+++ GameTypes/Cells/CharaCell.cs
@@ -10,12 +10,12 @@ namespace GameTypes.Cells
         /// <summary>
         /// Horizontal position [col]
         /// </summary>
-        public int X = 0;
+        public float X = 0;
 
         /// <summary>
         /// Vertical position [row]
         /// </summary>
-        public int Y = 0;
+        public float Y = 0;
 
         /// <summary>
         /// The chara type
